refactor render of chart
