+++ public/index.html
@@ -31,13 +31,6 @@
     </header>
     <main class="page-body__page-main  page-main">
       <div class="page-body__container">
-          <section class="trip-events">
-            <h2 class="visually-hidden">Trip events</h2>
-
-            <!-- Сортировка -->
-
-            <!-- Контент -->
-          </section>
       </div>
     </main>
     <script src="bundle.js"></script>
