+++ src/main.js
@@ -6,7 +6,7 @@ import FilterPresenter from "./presenter/filter.js";
 import InfoPresenter from "./presenter/info.js";
 import EventsModel from "./model/events.js";
 import FilterModel from "./model/filter.js";
-import {render, RenderPosition} from "./utils/render.js";
+import {render, remove, RenderPosition} from "./utils/render.js";
 import {sortEventDay} from "./utils/event.js";
 import {MenuItem} from "./const.js";
 
@@ -14,17 +14,19 @@ const EVENT_COUNT = 16;
 
 const siteControlsElement = document.querySelector(`.trip-controls`);
 const siteMainElement = document.querySelector(`.page-main .page-body__container`);
-const siteEventsElement = siteMainElement.querySelector(`.trip-events`);
+const addNewButtonElement = document.querySelector(`.trip-main__event-add-btn`);
 const [siteMenuHeader, siteFilterHeader] = siteControlsElement.querySelectorAll(`h2`);
 const events = new Array(EVENT_COUNT).fill().map(generateEvent).sort(sortEventDay);
 const siteMenuComponent = new SiteMenuView();
+
+let statisticsComponent = null;
 
 const eventsModel = new EventsModel();
 eventsModel.setEvents(events);
 
 const filterModel = new FilterModel();
 
-const tripPresenter = new TripPresenter(siteEventsElement, eventsModel, filterModel);
+const tripPresenter = new TripPresenter(siteMainElement, eventsModel, filterModel);
 const filterPresenter = new FilterPresenter(siteFilterHeader, filterModel, eventsModel);
 const tripInfoPresenter = new InfoPresenter(eventsModel.getEvents());
 
@@ -33,12 +35,19 @@ render(siteMenuHeader, siteMenuComponent, RenderPosition.AFTER);
 const handleSiteMenuClick = (menuItem) => {
   switch (menuItem) {
     case MenuItem.TABLE:
+      siteMenuComponent.setMenuItem(MenuItem.TABLE);
+      tripInfoPresenter.destroy();
       tripPresenter.init();
-      // Скрыть статистику
+      remove(statisticsComponent);
+      addNewButtonElement.disabled = false;
       break;
     case MenuItem.STATISTICS:
+      siteMenuComponent.setMenuItem(MenuItem.STATISTICS);
+      statisticsComponent = new StatisticsView(eventsModel.getEvents());
+      render(siteMainElement, statisticsComponent, RenderPosition.BEFOREEND);
+      tripInfoPresenter.init();
       tripPresenter.destroy();
-      // Показать статистику
+      addNewButtonElement.disabled = true;
       break;
   }
 };
@@ -46,11 +55,9 @@ const handleSiteMenuClick = (menuItem) => {
 siteMenuComponent.setMenuClickHandler(handleSiteMenuClick);
 
 filterPresenter.init();
-tripInfoPresenter.init();
-// tripPresenter.init();
-render(siteMainElement, new StatisticsView(eventsModel.getEvents()), RenderPosition.BEFOREEND);
+tripPresenter.init();
 
-document.querySelector(`.trip-main__event-add-btn`).addEventListener(`click`, (evt) => {
+addNewButtonElement.addEventListener(`click`, (evt) => {
   evt.preventDefault();
   tripPresenter.createEvent();
 });
