+++ src/view/statistics.js
@@ -1,232 +1,5 @@
 import AbstractView from "./abstract-view.js";
-import {makeItemsUniq, countMoneyByType, countEventsByType, countTimeByType} from "../utils/statistics.js";
-
-import Chart from "chart.js";
-import ChartDataLabels from 'chartjs-plugin-datalabels';
-
-const BAR_HEIGHT = 55;
-
-const renderMoneyChart = (moneyCtx, events) => {
-  const eventTypes = events.map((event) => event.type).sort();
-  const uniqEvents = makeItemsUniq(eventTypes);
-  const moneyByTypes = uniqEvents.map((event) => countMoneyByType(events, event));
-  const typeLabels = uniqEvents.map((event) => event.toUpperCase());
-
-  moneyCtx.height = BAR_HEIGHT * uniqEvents.length - 1;
-
-  return new Chart(moneyCtx, {
-    plugins: [ChartDataLabels],
-    type: `horizontalBar`,
-    data: {
-      labels: typeLabels,
-      datasets: [{
-        data: moneyByTypes,
-        backgroundColor: `#ffffff`,
-        hoverBackgroundColor: `#ffffff`,
-        anchor: `start`,
-        barThickness: 44,
-        minBarLength: 50
-      }]
-    },
-    options: {
-      plugins: {
-        datalabels: {
-          font: {
-            size: 13
-          },
-          color: `#000000`,
-          anchor: `end`,
-          align: `start`,
-          formatter: (val) => `€ ${val}`
-        }
-      },
-      title: {
-        display: true,
-        text: `MONEY`,
-        fontColor: `#000000`,
-        fontSize: 23,
-        position: `left`
-      },
-      scales: {
-        yAxes: [{
-          ticks: {
-            fontColor: `#000000`,
-            padding: 5,
-            fontSize: 13,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false
-          }
-        }],
-        xAxes: [{
-          ticks: {
-            display: false,
-            beginAtZero: true,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false
-          }
-        }],
-      },
-      legend: {
-        display: false
-      },
-      tooltips: {
-        enabled: false,
-      }
-    }
-  });
-};
-
-const renderTypeChart = (typeCtx, events) => {
-  const eventTypes = events.map((event) => event.type).sort();
-  const uniqEvents = makeItemsUniq(eventTypes);
-  const eventByTypes = uniqEvents.map((event) => countEventsByType(events, event));
-  const typeLabels = uniqEvents.map((event) => event.toUpperCase());
-
-  typeCtx.height = BAR_HEIGHT * uniqEvents.length - 1;
-
-  return new Chart(typeCtx, {
-    plugins: [ChartDataLabels],
-    type: `horizontalBar`,
-    data: {
-      labels: typeLabels,
-      datasets: [{
-        data: eventByTypes,
-        backgroundColor: `#ffffff`,
-        hoverBackgroundColor: `#ffffff`,
-        anchor: `start`,
-        barThickness: 44,
-        minBarLength: 50
-      }]
-    },
-    options: {
-      plugins: {
-        datalabels: {
-          font: {
-            size: 13
-          },
-          color: `#000000`,
-          anchor: `end`,
-          align: `start`,
-          formatter: (val) => `${val}x`
-        }
-      },
-      title: {
-        display: true,
-        text: `TYPE`,
-        fontColor: `#000000`,
-        fontSize: 23,
-        position: `left`
-      },
-      scales: {
-        yAxes: [{
-          ticks: {
-            fontColor: `#000000`,
-            padding: 5,
-            fontSize: 13,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false
-          }
-        }],
-        xAxes: [{
-          ticks: {
-            display: false,
-            beginAtZero: true,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false
-          }
-        }],
-      },
-      legend: {
-        display: false
-      },
-      tooltips: {
-        enabled: false,
-      }
-    }
-  });
-};
-
-const renderTimeChart = (timeCtx, events) => {
-  const eventTypes = events.map((event) => event.type).sort();
-  const uniqEvents = makeItemsUniq(eventTypes);
-  const timeByTypes = uniqEvents.map((event) => countTimeByType(events, event));
-  const typeLabels = uniqEvents.map((event) => event.toUpperCase());
-
-  timeCtx.height = BAR_HEIGHT * uniqEvents.length - 1;
-
-  return new Chart(timeCtx, {
-    plugins: [ChartDataLabels],
-    type: `horizontalBar`,
-    data: {
-      labels: typeLabels,
-      datasets: [{
-        data: timeByTypes,
-        backgroundColor: `#ffffff`,
-        hoverBackgroundColor: `#ffffff`,
-        anchor: `start`,
-        barThickness: 44,
-        minBarLength: 50
-      }]
-    },
-    options: {
-      plugins: {
-        datalabels: {
-          font: {
-            size: 13
-          },
-          color: `#000000`,
-          anchor: `end`,
-          align: `start`,
-          formatter: (val) => `${val}D`
-        }
-      },
-      title: {
-        display: true,
-        text: `TIME-SPEND`,
-        fontColor: `#000000`,
-        fontSize: 23,
-        position: `left`
-      },
-      scales: {
-        yAxes: [{
-          ticks: {
-            fontColor: `#000000`,
-            padding: 5,
-            fontSize: 13,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false
-          }
-        }],
-        xAxes: [{
-          ticks: {
-            display: false,
-            beginAtZero: true,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false
-          }
-        }],
-      },
-      legend: {
-        display: false
-      },
-      tooltips: {
-        enabled: false,
-      }
-    }
-  });
-};
+import {renderMoneyChart, renderTypeChart, renderTimeChart} from "../utils/statistics.js";
 
 const createStatisticsTemplate = () => {
 
@@ -284,10 +57,6 @@ export default class Statistics extends AbstractView {
     const moneyCtx = this.getElement().querySelector(`.statistics__chart--money`);
     const typeCtx = this.getElement().querySelector(`.statistics__chart--transport`);
     const timeCtx = this.getElement().querySelector(`.statistics__chart--time`);
-
-
-    typeCtx.height = BAR_HEIGHT * 5;
-    timeCtx.height = BAR_HEIGHT * 5;
 
     this._moneyChart = renderMoneyChart(moneyCtx, this._events);
     this._typeChart = renderTypeChart(typeCtx, this._events);
